study react 2022-01-09

+++ _study/nomadcoders/react_TwitterClone/src/components/Tweet.js
@@ -9,7 +9,9 @@ const Tweet = ({ tweetObj, isOwner }) => {
     const ok = window.confirm('Are you sure you want to delete this tweet?');
     if (ok) {
       await dbService.doc(`tweets/${tweetObj.id}`).delete();
-      await storageService.refFromURL(tweetObj.imageUrl).delete();
+      if (tweetObj.imageUrl) {
+        await storageService.refFromURL(tweetObj.imageUrl).delete();
+      }
     }
   };
   const toggleEditing = () => {
